feat: refactor array tool kit
