+++ utils/array-tool-kit.js
@@ -15,6 +15,18 @@ class ArrayToolKit {
     keyword = keyword.toLowerCase()
     return data.filter(item => item[field].toLowerCase() === keyword)
   }
+
+  // 產出不重複的索引值
+  static generateOptions(optionNum, optionMaxNum) {
+    const optionHashTab = {}
+    while (true) {
+      const selectedOption = Math.floor(Math.random() * optionMaxNum)
+      optionHashTab[`${selectedOption}`] = true
+      if (Object.keys(optionHashTab).length === optionNum) break
+    }
+
+    return Object.keys(optionHashTab)
+  }
 }
 
 exports = module.exports = {
